pyproject.toml
[tool.poetry]
name = "modularnotes"
version = "0.1.0"
description = ""
authors = ["J.Valkonen <johannes.valkonen@helsinki.fi>"]

[tool.poetry.dependencies]
python = "^3.6"
invoke = "^1.5.0"
pylint = "^2.7.4"
python-dotenv = "^0.16.0"
PyQt5 = "^5.15.4"
Faker = "^8.1.0"

[tool.poetry.dev-dependencies]
coverage = "^5.5"
pytest = "^6.2.2"
autopep8 = "^1.5.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
